Fix contributewd by Roland Dowdeswell for 64-bit bug in name canon patches

+++ lib/krb5/principal.c
@@ -1477,8 +1477,8 @@ parse_name_canon_rules(krb5_context context, char **rulestrs,
     char *tok;
     char *cp;
     char **cpp;
-    unsigned int n = 0;
-    unsigned int i, k;
+    size_t n = 0;
+    size_t i, k;
     krb5_name_canon_rule r;
 
     for (cpp = rulestrs; *cpp; cpp++)
